moved integration tests to p7 or broken
